added the entity extraction and orphan node deletion in tabs
