+++ frontend/src/components/Layout/PageLayout.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 import SideNav from './SideNav';
 import DrawerDropzone from './DrawerDropzone';
 import DrawerChatbot from './DrawerChatbot';
@@ -13,23 +13,15 @@ import { OverridableStringUnion } from '@mui/types';
 import { useFileContext } from '../../context/UsersFiles';
 import SchemaFromTextDialog from '../Popups/Settings/SchemaFromText';
 import CustomAlert from '../UI/Alert';
-import DeletePopUpForOrphanNodes from '../Popups/DeletePopUpForOrphanNodes';
-import deleteOrphanAPI from '../../services/DeleteOrphanNodes';
 
 export default function PageLayoutNew({
   isSettingPanelExpanded,
   closeSettingModal,
   openSettingsDialog,
-  closeOrphanNodeDeletionModal,
-  showOrphanNodeDeletionModal,
-  openOrphanNodeDeletionModal,
 }: {
   isSettingPanelExpanded: boolean;
   closeSettingModal: () => void;
   openSettingsDialog: () => void;
-  closeOrphanNodeDeletionModal: () => void;
-  showOrphanNodeDeletionModal: boolean;
-  openOrphanNodeDeletionModal: () => void;
 }) {
   const [isLeftExpanded, setIsLeftExpanded] = useState<boolean>(true);
   const [isRightExpanded, setIsRightExpanded] = useState<boolean>(false);
@@ -39,17 +31,13 @@ export default function PageLayoutNew({
   const { userCredentials } = useCredentials();
   const toggleLeftDrawer = () => setIsLeftExpanded(!isLeftExpanded);
   const toggleRightDrawer = () => setIsRightExpanded(!isRightExpanded);
-  const [openTextSchemaDialog, setOpenTextSchemaDialog] = useState<boolean>(false);
-  const [orphanDeleteAPIloading, setorphanDeleteAPIloading] = useState<boolean>(false);
   const [alertDetails, setalertDetails] = useState<alertStateType>({
     showAlert: false,
     alertType: 'error',
     alertMessage: '',
   });
   const { messages } = useMessageContext();
-  const openSchemaFromTextDialog = useCallback(() => setOpenTextSchemaDialog(true), []);
-  const closeSchemaFromTextDialog = useCallback(() => setOpenTextSchemaDialog(false), []);
-  const { isSchema, setIsSchema } = useFileContext();
+  const { isSchema, setIsSchema, setShowTextFromSchemaDialog, showTextFromSchemaDialog } = useFileContext();
 
   const deleteOnClick = async () => {
     try {
@@ -84,16 +72,7 @@ export default function PageLayoutNew({
       alertMessage: '',
     });
   };
-  const orphanNodesDeleteHandler = async (selectedEntities: string[]) => {
-    try {
-      setorphanDeleteAPIloading(true);
-      const response = await deleteOrphanAPI(userCredentials as UserCredentials, selectedEntities);
-      setorphanDeleteAPIloading(false);
-      console.log(response);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+
   return (
     <div style={{ maxHeight: 'calc(100vh - 58px)' }} className='flex overflow-hidden'>
       {alertDetails.showAlert && (
@@ -107,19 +86,17 @@ export default function PageLayoutNew({
       <SideNav isExpanded={isLeftExpanded} position='left' toggleDrawer={toggleLeftDrawer} />
       <DrawerDropzone isExpanded={isLeftExpanded} />
       <SchemaFromTextDialog
-        open={openTextSchemaDialog}
+        open={showTextFromSchemaDialog}
         openSettingsDialog={openSettingsDialog}
-        onClose={closeSchemaFromTextDialog}
+        onClose={() => {
+          setShowTextFromSchemaDialog(false);
+        }}
         showAlert={showAlert}
       ></SchemaFromTextDialog>
-      <DeletePopUpForOrphanNodes
-        open={showOrphanNodeDeletionModal}
-        deleteCloseHandler={closeOrphanNodeDeletionModal}
-        deleteHandler={orphanNodesDeleteHandler}
-        loading={orphanDeleteAPIloading}
-      ></DeletePopUpForOrphanNodes>
       <SettingsModal
-        openTextSchema={openSchemaFromTextDialog}
+        openTextSchema={() => {
+          setShowTextFromSchemaDialog(true);
+        }}
         open={isSettingPanelExpanded}
         onClose={closeSettingModal}
         settingView='headerView'
@@ -131,10 +108,12 @@ export default function PageLayoutNew({
         isLeftExpanded={isLeftExpanded}
         isRightExpanded={isRightExpanded}
         showChatBot={showChatBot}
-        openTextSchema={openSchemaFromTextDialog}
-        openOrphanNodeDeletionModal={openOrphanNodeDeletionModal}
+        openTextSchema={() => {
+          setShowTextFromSchemaDialog(true);
+        }}
         isSchema={isSchema}
         setIsSchema={setIsSchema}
+        openSettingsDialog={openSettingsDialog}
       />
       {showDrawerChatbot && (
         <DrawerChatbot messages={messages} isExpanded={isRightExpanded} clearHistoryData={clearHistoryData} />
