+++ frontend/src/components/Popups/GraphEnhancementDialog/index.tsx
@@ -1,7 +1,40 @@
-import { Dialog, Tabs } from '@neo4j-ndl/react';
+import { Dialog, Tabs, Box, Typography } from '@neo4j-ndl/react';
+import graphenhancement from '../../../assets/images/graph-enhancements.svg';
 import { useState } from 'react';
+import DeletePopUpForOrphanNodes from '../DeletePopUpForOrphanNodes';
+import deleteOrphanAPI from '../../../services/DeleteOrphanNodes';
+import { UserCredentials } from '../../../types';
+import { useCredentials } from '../../../context/UserCredentials';
+import EntityExtractionSettings from '../Settings/EntityExtractionSetting';
+import { AlertColor, AlertPropsColorOverrides } from '@mui/material';
+import { OverridableStringUnion } from '@mui/types';
+import { useFileContext } from '../../../context/UsersFiles';
 
-export default function GraphEnhancementDialog({ open, onClose }: { open: boolean; onClose: () => void }) {
+export default function GraphEnhancementDialog({
+  open,
+  onClose,
+}: {
+  open: boolean;
+  onClose: () => void;
+  showAlert: (
+    alertmsg: string,
+    alerttype: OverridableStringUnion<AlertColor, AlertPropsColorOverrides> | undefined
+  ) => void;
+}) {
+  const [orphanDeleteAPIloading, setorphanDeleteAPIloading] = useState<boolean>(false);
+  const { setShowTextFromSchemaDialog } = useFileContext();
+  const { userCredentials } = useCredentials();
+
+  const orphanNodesDeleteHandler = async (selectedEntities: string[]) => {
+    try {
+      setorphanDeleteAPIloading(true);
+      const response = await deleteOrphanAPI(userCredentials as UserCredentials, selectedEntities);
+      setorphanDeleteAPIloading(false);
+      console.log(response);
+    } catch (error) {
+      console.log(error);
+    }
+  };
   const [activeTab, setactiveTab] = useState<number>(0);
   return (
     <Dialog
@@ -15,7 +48,19 @@ export default function GraphEnhancementDialog({ open, onClose }: { open: boolea
       onClose={onClose}
     >
       <Dialog.Header className='flex justify-between self-end' id='chatbot-dialog-title'>
-        Graph Enhancement Operations
+        <Box className='n-bg-palette-neutral-bg-weak p-4'>
+          <Box className='flex flex-row pb-6 items-center mb-2'>
+            <img src={graphenhancement} style={{ width: 95, height: 95, marginRight: 10 }} loading='lazy' />
+            <Box className='flex flex-col'>
+              <Typography variant='h2'>Graph Enhancement Operations</Typography>
+              <Typography variant='body-medium' className='mb-2'>
+                This set of tools will help you enhance the quality of your Knowledge Graph by removing possible
+                duplicated entities, disconnected nodes and set a Graph Schema for improving the quality of the entity
+                extraction process
+              </Typography>
+            </Box>
+          </Box>
+        </Box>
       </Dialog.Header>
       <Dialog.Content className='flex flex-col n-gap-token-4 w-full grow overflow-auto'>
         <Tabs fill='underline' onChange={setactiveTab} size='large' value={activeTab}>
@@ -27,10 +72,19 @@ export default function GraphEnhancementDialog({ open, onClose }: { open: boolea
           </Tabs.Tab>
         </Tabs>
         <Tabs.TabPanel className='n-flex n-flex-col n-gap-token-4 n-p-token-6' value={activeTab} tabId={0}>
-          <div></div>
+          <>
+            <EntityExtractionSettings
+              view='Tabs'
+              openTextSchema={() => {
+                setShowTextFromSchemaDialog(true);
+              }}
+              colseEnhanceGraphSchemaDialog={onClose}
+              settingView='headerView'
+            ></EntityExtractionSettings>
+          </>
         </Tabs.TabPanel>
         <Tabs.TabPanel className='n-flex n-flex-col n-gap-token-4 n-p-token-6' value={activeTab} tabId={1}>
-          <div></div>
+          <DeletePopUpForOrphanNodes deleteHandler={orphanNodesDeleteHandler} loading={orphanDeleteAPIloading} />
         </Tabs.TabPanel>
       </Dialog.Content>
     </Dialog>
