+++ frontend/src/types.ts
@@ -153,7 +153,7 @@ export interface ContentProps {
   openTextSchema: () => void;
   isSchema?: boolean;
   setIsSchema: Dispatch<SetStateAction<boolean>>;
-  openOrphanNodeDeletionModal: () => void;
+  openSettingsDialog: () => void;
 }
 
 export interface FileTableProps {
